game logic: draw possible moves and fixing chessboard

+++ Chess/OpenGL-Core/src/Utilities/DeltaTime.h
@@ -9,10 +9,9 @@ public:
 
 	inline float GetMilliSeconds() const { return m_FrameTime * 1000.0f; }
 	inline float GetSeconds() const { return m_FrameTime; }
-	inline int GetFramePerSecounds() const { return (int) (1.0f / m_FrameTime); }
+	inline int GetFramePerSecounds() const { return (int)(1.0f / m_FrameTime); }
 
 private:
-	float m_LastFrameTime = 0.0f;
+	double m_LastFrameTime = 0.0f;
 	float m_FrameTime = 0.0f;
-
 };
